remove proxy layer, inline HTTP logic into ShoppingListService
