+++ WinUI/Views/ShoppingListView.xaml.cs
@@ -12,8 +12,7 @@ public sealed partial class ShoppingListView : Page
     public ShoppingListView()
     {
         this.InitializeComponent();
-        var proxy = new ShoppingListServiceProxy(new HttpClient());
-        this.ViewModel = new ShoppingListViewModel(new ShoppingListService(proxy), new UserSession());
+        this.ViewModel = new ShoppingListViewModel(new ShoppingListService(new HttpClient()), new UserSession());
         this.DataContext = this.ViewModel;
         this.Loaded += this.OnLoaded;
     }
